get all stack accesses to pass type checks
This is a prerequisite for using a memory optimized runtime stack. Also,
eliminate dead operations (no more maxInt / 4 hacks!), which is probably
the source of decreased memory usage, and adds the necessary opcodes for
stack accesses to use the correct types.  However, the running time took
a big hit, which is an expected regression until more opcodes are added.

8m09s, 627 MB RSS